requirements.txt
streamlit
torch
torchvision
timm
numpy
Pillow
matplotlib
scikit-learn
tabulate
